@@ -1,51 +1,50 @@
 source 'https://rubygems.org'
 
-gem 'rails', '3.2.7'
+gem 'rails', '4.1.6'
 
-gem 'devise', '~>2.1.0'
-gem 'cancan', '~>1.6.7'
+# gem 'devise', '~>2.1.0'
+# gem 'cancan', '~>1.6.7'
 gem 'sqlite3-ruby' 
-gem 'delorean'
+# gem 'delorean'
 gem 'inherited_resources', '~> 1.3.0'
 gem 'haml', '~> 3.1.4'
 gem 'formtastic', '~> 2.2.0'
-gem 'cocoon', '~> 1.0.20'
-gem 'flexible_date', '~> 0.4.1'
-gem 'state_machine', '~> 1.1.2'
-gem 'rufus-scheduler'
-gem "ckeditor", "~> 3.7.0"
-gem "paperclip"
+# gem 'cocoon', '~> 1.0.20'
+# gem 'flexible_date', '~> 0.4.1'
+# gem 'state_machine', '~> 1.1.2'
+# gem 'rufus-scheduler'
+# gem "ckeditor", "~> 3.7.0"
+# gem "paperclip"
 gem 'jquery-rails'
 gem 'rabl'
 gem 'gon'
 gem 'multi_json', '~>1.3.0'
 gem 'erb2haml'
-gem 'nokogiri'
-gem 'mime-types'
+# gem 'nokogiri'
+# gem 'mime-types'
 
 group :assets do
-  gem 'sass-rails',   '~> 3.2.3'
-  gem 'coffee-rails', '~> 3.2.1'
-  gem 'therubyracer'
+  # gem 'sass-rails',   '~> 3.2.3'
+  # gem 'coffee-rails', '~> 3.2.1'
+  # gem 'therubyracer'
   gem 'uglifier', '>= 1.0.3'
 end
 
 group :test, :development do
-  gem 'sqlite3'
-  gem 'rspec-rails', '~> 2.11.0'
-  gem 'capybara', '~> 1.1.2'
-  gem 'capybara-webkit', '~> 0.12.0'
-  gem 'poltergeist', '~> 0.7.0'
-  gem 'launchy', '~> 2.1.0'
-  gem 'factory_girl_rails', '~> 3.6.0'
-  gem 'valid_attribute', '~> 1.2.0'
-  gem 'timecop', '~> 0.4.4'
-  gem 'pry'
-  gem 'pry-rails'
-  gem 'capistrano'
-  gem 'spork-rails', '~> 3.2.0'
-  gem 'simplecov', require: false
-  gem 'simplecov-rcov', require: false
-  gem 'zippy'
-  gem 'toothbrush'
+  # gem 'sqlite3'
+  # gem 'rspec-rails', '~> 2.11.0'
+  # gem 'capybara', '~> 1.1.2'
+  # gem 'capybara-webkit', '~> 0.12.0'
+  # gem 'poltergeist', '~> 0.7.0'
+  # gem 'launchy', '~> 2.1.0'
+  # gem 'factory_girl_rails', '~> 3.6.0'
+  # gem 'valid_attribute', '~> 1.2.0'
+  # gem 'timecop', '~> 0.4.4'
+  # gem 'pry'
+  # gem 'pry-rails'
+  # gem 'simplecov', require: false
+  # gem 'simplecov-rcov', require: false
+  # gem 'zippy'
+  # gem 'toothbrush'
+  gem 'dinde', :git => 'git@gitlab.com:fnoratol/dinde.git'
 end
