@@ -9,76 +9,76 @@
 # from scratch. The latter is a flawed and unsustainable approach (the more migrations
 # you'll amass, the slower it'll run and the greater likelihood for issues).
 #
-# It's strongly recommended to check this file into your version control system.
+# It's strongly recommended that you check this file into your version control system.
 
-ActiveRecord::Schema.define(:version => 20130421185805) do
+ActiveRecord::Schema.define(version: 20130421185805) do
 
-  create_table "eventos", :force => true do |t|
+  create_table "eventos", force: true do |t|
     t.string   "nome"
     t.string   "data"
-    t.datetime "created_at",     :null => false
-    t.datetime "updated_at",     :null => false
+    t.datetime "created_at"
+    t.datetime "updated_at"
     t.integer  "local_id"
     t.integer  "organizador_id"
     t.integer  "repertorio_id"
   end
 
-  create_table "fontes", :force => true do |t|
+  create_table "fontes", force: true do |t|
     t.string   "nome"
     t.integer  "corrente"
     t.integer  "mala_id"
-    t.datetime "created_at",  :null => false
-    t.datetime "updated_at",  :null => false
+    t.datetime "created_at"
+    t.datetime "updated_at"
     t.integer  "comprimento"
     t.integer  "largura"
   end
 
-  create_table "locais", :force => true do |t|
+  create_table "locais", force: true do |t|
     t.string   "nome"
     t.string   "endereco"
     t.string   "cidade"
     t.string   "tipo"
-    t.datetime "created_at",      :null => false
-    t.datetime "updated_at",      :null => false
+    t.datetime "created_at"
+    t.datetime "updated_at"
     t.integer  "proprietario_id"
   end
 
-  create_table "malas", :force => true do |t|
+  create_table "malas", force: true do |t|
     t.string   "nome"
     t.integer  "largura"
     t.integer  "comprimento"
-    t.datetime "created_at",  :null => false
-    t.datetime "updated_at",  :null => false
+    t.datetime "created_at"
+    t.datetime "updated_at"
   end
 
-  create_table "musicas", :force => true do |t|
+  create_table "musicas", force: true do |t|
     t.string   "nome"
     t.integer  "duracao"
     t.integer  "bpm"
     t.string   "estilo"
-    t.datetime "created_at", :null => false
-    t.datetime "updated_at", :null => false
+    t.datetime "created_at"
+    t.datetime "updated_at"
   end
 
-  create_table "ordemmusicas", :force => true do |t|
+  create_table "ordemmusicas", force: true do |t|
     t.integer  "ordem"
     t.integer  "repertorio_id"
     t.integer  "musica_id"
-    t.datetime "created_at",    :null => false
-    t.datetime "updated_at",    :null => false
+    t.datetime "created_at"
+    t.datetime "updated_at"
   end
 
-  add_index "ordemmusicas", ["musica_id"], :name => "index_ordemmusicas_on_musica_id"
-  add_index "ordemmusicas", ["repertorio_id"], :name => "index_ordemmusicas_on_repertorio_id"
+  add_index "ordemmusicas", ["musica_id"], name: "index_ordemmusicas_on_musica_id"
+  add_index "ordemmusicas", ["repertorio_id"], name: "index_ordemmusicas_on_repertorio_id"
 
-  create_table "organizadors", :force => true do |t|
+  create_table "organizadors", force: true do |t|
     t.string   "nome"
     t.string   "contato"
-    t.datetime "created_at", :null => false
-    t.datetime "updated_at", :null => false
+    t.datetime "created_at"
+    t.datetime "updated_at"
   end
 
-  create_table "pedais", :force => true do |t|
+  create_table "pedais", force: true do |t|
     t.string   "nome"
     t.integer  "largura"
     t.integer  "comprimento"
@@ -86,30 +86,30 @@ ActiveRecord::Schema.define(:version => 20130421185805) do
     t.integer  "corrente"
     t.string   "tipo"
     t.integer  "mala_id"
-    t.datetime "created_at",  :null => false
-    t.datetime "updated_at",  :null => false
+    t.datetime "created_at"
+    t.datetime "updated_at"
   end
 
-  create_table "proprietarios", :force => true do |t|
+  create_table "proprietarios", force: true do |t|
     t.string   "nome"
     t.string   "contato"
-    t.datetime "created_at", :null => false
-    t.datetime "updated_at", :null => false
+    t.datetime "created_at"
+    t.datetime "updated_at"
   end
 
-  create_table "repertorios", :force => true do |t|
+  create_table "repertorios", force: true do |t|
     t.string   "nome"
-    t.datetime "created_at", :null => false
-    t.datetime "updated_at", :null => false
+    t.datetime "created_at"
+    t.datetime "updated_at"
   end
 
-  create_table "saida_fontes", :force => true do |t|
-    t.integer  "tensao",          :default => 9
-    t.boolean  "centro_negativo", :default => true
+  create_table "saida_fontes", force: true do |t|
+    t.integer  "tensao",          default: 9
+    t.boolean  "centro_negativo", default: true
     t.integer  "pedal_id"
     t.integer  "fonte_id"
-    t.datetime "created_at",                        :null => false
-    t.datetime "updated_at",                        :null => false
+    t.datetime "created_at"
+    t.datetime "updated_at"
   end
 
 end
