@@ -1,5 +1,5 @@
 class Local < ActiveRecord::Base
-  attr_accessible :cidade, :endereco, :nome, :tipo, :proprietario_id
+  # attr_accessible :cidade, :endereco, :nome, :tipo, :proprietario_id
   belongs_to :proprietario
   has_many :eventos
 
