@@ -1,4 +1,4 @@
 class Organizador < ActiveRecord::Base
-  attr_accessible :contato, :nome
+  # attr_accessible :contato, :nome
   has_many :eventos
 end
