@@ -1,4 +1,4 @@
 class Proprietario < ActiveRecord::Base
-  attr_accessible :contato, :nome
+  # attr_accessible :contato, :nome
   has_many :locais
 end
