@@ -1,6 +1,6 @@
 class Repertorio < ActiveRecord::Base
-  attr_accessible :nome
-	has_many :ordemmusicas
+  # attr_accessible :nome
+  has_many :ordemmusicas
   has_many :musicas, :through => :ordemmusicas
   has_many :eventos
 end
